test: move tests next to code
